# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Voice as a Service",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/tiers": {
   "get": {
    "tags": [
     "tiers"
    ],
    "summary": "List Tiers",
    "operationId": "list_tiers_api_v1_tiers_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/TierOut"
         },
         "type": "array",
         "title": "Response List Tiers Api V1 Tiers Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "post": {
    "tags": [
     "tiers"
    ],
    "summary": "Create Tier",
    "operationId": "create_tier_api_v1_tiers_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TierIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TierOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/tiers/{tier_id}": {
   "get": {
    "tags": [
     "tiers"
    ],
    "summary": "Get Tier",
    "operationId": "get_tier_api_v1_tiers__tier_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "tier_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Tier Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TierOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "tiers"
    ],
    "summary": "Update Tier",
    "description": "Solo los campos enviados; un limite en null lo deja ilimitado (la duracion vuelve a\nCALL_MAX_DURATION_SECONDS; la retencion, sin borrado). Aplica desde ya al mes en curso.",
    "operationId": "update_tier_api_v1_tiers__tier_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "tier_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Tier Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TierUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TierOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "tiers"
    ],
    "summary": "Delete Tier",
    "operationId": "delete_tier_api_v1_tiers__tier_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "tier_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Tier Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/phone-numbers": {
   "get": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "List Numbers",
    "description": "Admin: todo el inventario (status=free: libres). Cliente: los suyos.",
    "operationId": "list_numbers_api_v1_phone_numbers_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "client_id",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Client Id"
      }
     },
     {
      "name": "status",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "enum": [
          "free",
          "assigned"
         ],
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Status"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/PhoneNumberOut"
         },
         "title": "Response List Numbers Api V1 Phone Numbers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Create Number",
    "description": "Carga un numero; con client_id lo asigna (tope del tier) y con agent_id lo rutea.",
    "operationId": "create_number_api_v1_phone_numbers_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneNumberIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneNumberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/phone-numbers/bulk": {
   "post": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Load Numbers",
    "description": "Carga varios numeros libres al inventario; los repetidos o invalidos se informan en skipped.",
    "operationId": "load_numbers_api_v1_phone_numbers_bulk_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneNumberBulkIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneNumberBulkOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/phone-numbers/{number_id}/assign": {
   "post": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Assign Number",
    "description": "Asigna un numero libre a un cliente. 409 phone_numbers_limit si su tier no tiene lugar.",
    "operationId": "assign_number_api_v1_phone_numbers__number_id__assign_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "number_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Number Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AssignIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneNumberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/phone-numbers/{number_id}/release": {
   "post": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Release Number",
    "description": "Lo devuelve al inventario (sin cliente ni agente): sus entrantes dejan de atenderse.",
    "operationId": "release_number_api_v1_phone_numbers__number_id__release_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "number_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Number Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneNumberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/phone-numbers/{number_id}": {
   "patch": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Update Number",
    "description": "Etiqueta y agente que atiende las entrantes (un agente del mismo cliente). Cliente\ninactivo: 403 client_inactive (solo lectura, H12).",
    "operationId": "update_number_api_v1_phone_numbers__number_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "number_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Number Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PhoneNumberUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PhoneNumberOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "phone-numbers"
    ],
    "summary": "Delete Number",
    "description": "Solo libres: uno asignado se libera primero.",
    "operationId": "delete_number_api_v1_phone_numbers__number_id__delete",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "number_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Number Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AssignIn": {
    "properties": {
     "client_id": {
      "type": "string",
      "title": "Client Id"
     }
    },
    "type": "object",
    "required": [
     "client_id"
    ],
    "title": "AssignIn"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "PhoneNumberBulkIn": {
    "properties": {
     "numbers": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "maxItems": 1000,
      "minItems": 1,
      "title": "Numbers"
     },
     "label": {
      "type": "string",
      "maxLength": 64,
      "title": "Label",
      "default": ""
     },
     "provider": {
      "type": "string",
      "maxLength": 32,
      "title": "Provider",
      "default": "anura"
     }
    },
    "type": "object",
    "required": [
     "numbers"
    ],
    "title": "PhoneNumberBulkIn",
    "description": "Carga al inventario: uno por elemento, en E.164 o con espacios y guiones."
   },
   "PhoneNumberBulkOut": {
    "properties": {
     "created": {
      "items": {
       "$ref": "#/components/schemas/PhoneNumberOut"
      },
      "type": "array",
      "title": "Created"
     },
     "skipped": {
      "items": {
       "$ref": "#/components/schemas/SkippedNumber"
      },
      "type": "array",
      "title": "Skipped"
     }
    },
    "type": "object",
    "required": [
     "created",
     "skipped"
    ],
    "title": "PhoneNumberBulkOut"
   },
   "PhoneNumberIn": {
    "properties": {
     "e164": {
      "type": "string",
      "pattern": "^\\+\\d{8,15}$",
      "title": "E164",
      "examples": [
       "+541152630861"
      ]
     },
     "label": {
      "type": "string",
      "maxLength": 64,
      "title": "Label",
      "default": ""
     },
     "provider": {
      "type": "string",
      "maxLength": 32,
      "title": "Provider",
      "default": "anura"
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id",
      "description": "Agente que atiende las entrantes"
     }
    },
    "type": "object",
    "required": [
     "e164"
    ],
    "title": "PhoneNumberIn"
   },
   "PhoneNumberOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "e164": {
      "type": "string",
      "title": "E164"
     },
     "label": {
      "type": "string",
      "title": "Label"
     },
     "provider": {
      "type": "string",
      "title": "Provider"
     },
     "client_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Id"
     },
     "client_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Name"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     },
     "agent_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Name"
     },
     "assigned_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Assigned At"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     }
    },
    "type": "object",
    "required": [
     "id",
     "e164",
     "label",
     "provider",
     "client_id",
     "agent_id",
     "assigned_at",
     "created_at"
    ],
    "title": "PhoneNumberOut"
   },
   "PhoneNumberUpdate": {
    "properties": {
     "label": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64
       },
       {
        "type": "null"
       }
      ],
      "title": "Label"
     },
     "agent_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Agent Id"
     }
    },
    "type": "object",
    "title": "PhoneNumberUpdate"
   },
...
   "SkippedNumber": {
    "properties": {
     "number": {
      "type": "string",
      "title": "Number"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     }
    },
    "type": "object",
    "required": [
     "number",
     "reason"
    ],
    "title": "SkippedNumber"
   },
...
   "TierIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 64,
      "minLength": 1,
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description",
      "default": ""
     },
     "max_concurrent_calls": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Concurrent Calls"
     },
     "inbound_minutes": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Inbound Minutes",
      "description": "Minutos entrantes por mes"
     },
     "outbound_minutes": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Outbound Minutes",
      "description": "Minutos salientes por mes"
     },
     "max_phone_numbers": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Phone Numbers",
      "description": "Numeros que puede tener el cliente"
     },
     "max_call_duration_seconds": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Call Duration Seconds",
      "description": "Tope duro de cada llamada, en segundos (null: el del sistema)"
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days",
      "description": "Dias que se guardan las conversaciones (null: sin borrado)"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "TierIn"
   },
   "TierOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "description": {
      "type": "string",
      "title": "Description"
     },
     "max_concurrent_calls": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Concurrent Calls"
     },
     "inbound_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Inbound Minutes"
     },
     "outbound_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Outbound Minutes"
     },
     "max_phone_numbers": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Phone Numbers"
     },
     "max_call_duration_seconds": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Call Duration Seconds"
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "clients_count": {
      "type": "integer",
      "title": "Clients Count",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "description",
     "max_concurrent_calls",
     "inbound_minutes",
     "outbound_minutes",
     "max_phone_numbers",
     "max_call_duration_seconds",
     "retention_days",
     "created_at"
    ],
    "title": "TierOut"
   },
   "TierUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 64,
        "minLength": 1
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "description": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Description"
     },
     "max_concurrent_calls": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Concurrent Calls"
     },
     "inbound_minutes": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Inbound Minutes"
     },
     "outbound_minutes": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Outbound Minutes"
     },
     "max_phone_numbers": {
      "anyOf": [
       {
        "type": "integer",
        "minimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Phone Numbers"
     },
     "max_call_duration_seconds": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Call Duration Seconds"
     },
     "retention_days": {
      "anyOf": [
       {
        "type": "integer",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Retention Days"
     }
    },
    "type": "object",
    "title": "TierUpdate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "description": "Token de la sesion de demo o de la llamada",
    "scheme": "bearer"
   }
  }
 }
}